factored out individual parsers from vcdrom
