+++ lib/vcdrom.js
@@ -5,12 +5,16 @@ const createVCD = require('vcd-stream/out/vcd.js');
 const webVcdParser = require('vcd-stream/lib/web-vcd-parser.js');
 const stringify = require('onml/stringify.js');
 const domContainer = require('@wavedrom/doppler/lib/dom-container.js');
+const renderValues = require('@wavedrom/doppler/lib/render-values.js');
+const defaultKeyBindo = require('@wavedrom/doppler/lib/key-bindo.js');
+
 const getReaders = require('./get-readers.js');
 const vcdPipeDeso = require('./vcd-pipe-deso.js');
 const dropZone = require('./drop-zone.js');
-const lister = require('./lister.js');
-
-const maxChunkLength = 1 << 22; // Number.MAX_SAFE_INTEGER; // 5e6; // 300000; // 1 << 23;
+const getWaveql = require('./get-waveql.js');
+const getListing = require('./get-listing.js');
+const getVcd = require('./get-vcd.js');
+const getJsonls = require('./get-jsonls.js');
 
 // const u8toStr = (u8) => {
 //   let res = '';
@@ -31,126 +35,6 @@ const getElement = divName => {
   return divName;
 };
 
-
-const getWaveql = async (readers) => {
-  let waveql;
-  const r = readers.find(reader => reader.ext === 'waveql');
-  if (r && (r.reader)) {
-    // console.log('WaveQL', r);
-    const utf8Decoder = new TextDecoder('utf-8');
-    waveql = '';
-    for (let i = 0; i < 1e5; i++) {
-      const { done, value } = await r.reader.read();
-      waveql += value ? utf8Decoder.decode(value, {stream: true}) : '';
-      if (done) {
-        break;
-      }
-    }
-  }
-  return waveql;
-};
-
-const getListing = async (readers) => {
-  let listing = [];
-  const r = readers.find(reader => reader.ext === 'lst');
-  if (r) {
-    // console.log('LST', r);
-    const utf8Decoder = new TextDecoder('utf-8');
-    const list = lister();
-    for (let i = 0; i < 10; i++) {
-      const { done, value } = await r.reader.read();
-      list.onChunk(value ? utf8Decoder.decode(value, {stream: true}) : '');
-      if (done) {
-        listing = list.getTrace();
-        break;
-      }
-    }
-  }
-  return listing;
-};
-
-// JSONL / NDJSON log files
-const getJsonls = async (readers) => {
-  const jsonls = [];
-  const utf8Decoder = new TextDecoder('utf-8');
-  for (const r of readers) {
-    if ((r.ext !== 'jsonl') && (r.ext !== 'jsonl')) {
-      continue;
-    }
-    // console.log('JSONL', r);
-    let tail = '';
-    const data = r.data = [];
-    let lineNumber = 0;
-    for (let i = 0; i < 1e5; i++) {
-      const { done, value } = await r.reader.read();
-      tail += value ? utf8Decoder.decode(value) : '';
-      const lines = tail.split(/\n/);
-      // console.log(i, lines.length);
-      for (let j = 0; j < (lines.length - 1); j++) {
-        lineNumber++;
-        try {
-          data.push(JSON.parse(lines[j]));
-        } catch (err) {
-          console.log('line: ' + lineNumber + ' chunk:' + i, lines[j], err);
-        }
-        tail = lines[lines.length - 1];
-      }
-      if (done) {
-        if (tail === '') {
-          break;
-        }
-        try {
-          data.push(JSON.parse(tail));
-        } catch (err) {
-          console.log(i, 'tail', err);
-        }
-        break;
-      }
-    }
-    jsonls.push(r);
-  }
-  // console.log(jsonls);
-  return jsonls;
-};
-
-const getVcd = async (readers, content, inst) => {
-  const r = readers.find(reader => reader.ext === 'vcd');
-  if (r) {
-    // console.log('VCD', r);
-    document.title = r.baseName;
-    content.innerHTML = '<div class="wd-progress">LOADING...</div>';
-    let total = 0;
-    outerLoop:
-    for (let i = 0; i < 1e5; i++) {
-      const { done, value } = await r.reader.read();
-
-      if (done && (value === undefined)) {
-        // console.log('the end');
-        inst.end();
-        break outerLoop;
-      }
-      const len = value.length;
-      for (let j = 0; j < len; j += maxChunkLength) {
-        const value1 = value.slice(j, j + maxChunkLength);
-        const len1 = value1.length;
-        total += len1;
-
-        // const vh = u8toStr(value1.slice(0, 100));
-        // const vt = u8toStr(value1.slice(-100));
-        // console.log({len1, done, total, vh, vt});
-
-        content.innerHTML = '<div class="wd-progress">' + total.toLocaleString() + '</div>';
-        if (done && ((j + maxChunkLength) >= len)) {
-          console.log('last chunk');
-          inst.end(value1);
-          break outerLoop;
-        }
-        inst.write(value1);
-      }
-    }
-  }
-};
-
 const getHandler = (content, inst) => async readers => {
   // console.log(readers);
 
@@ -167,7 +51,31 @@ const getHandler = (content, inst) => async readers => {
     deso.listing = listing;
     deso.timeOpt = timeOpt;
     deso.jsonls = jsonls;
-    domContainer(content, deso);
+
+    const container = domContainer();
+
+    // container.pstate.mySecret = 42;
+    // container.elemento
+    // container.layers
+    // container.renderPlugins.push(...)
+
+    Object.assign(container.keyBindo, defaultKeyBindo /* , additionalKeyBindings */ );
+
+    container.renderPlugins.push(
+      (desc, pstate, els) => {
+        const gen = renderValues(desc, pstate);
+        for (let i = 0; i < 1e6; i++) {
+          const iter = gen.next();
+          if (iter.done) {
+            els.values.innerHTML = iter.value;
+            break;
+          }
+        }
+      }
+    );
+
+    /* const {render} = */ container.start(content, deso);
+
   });
 
   await getVcd(readers, content, inst);
